requirements.txt
langchain>=0.2.0
langgraph>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pillow>=10.0.0
matplotlib>=3.7.0
